feat: extract stripe provider
